@luizperes 2020-09-07
Thanks for your pull request, @ChrisSmith! Will publish a new version soon.